fix: kpi BigDecimal로 수정

+++ src/main/java/org/backend/domain/analysis/dto/RfmKpiResponseDto.java
@@ -1,11 +1,13 @@
 package org.backend.domain.analysis.dto;
 
+import java.math.BigDecimal;
+
 public record RfmKpiResponseDto(
     String baseMonth,
-    Float crr,
+    BigDecimal crr,
     String crrStatus,
-    Float churnRate,
+    BigDecimal churnRate,
     String churnStatus,
-    Float nrr,
+    BigDecimal nrr,
     String nrrStatus
 ) {}
